requirements.txt
flask
python-dotenv
psycopg2-binary
authlib
requests
flask_cors
PyJWT 
pyopenssl
gunicorn
eventlet
DB
SQLAlchemy
Flask-Migrate